# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/ask":{"post":{"summary":"Ask","operationId":"ask_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AskRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/voice/save":{"post":{"summary":"Save Voice","operationId":"save_voice_voice_save_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveVoiceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/voice/process":{"post":{"summary":"Process Voice","operationId":"process_voice_voice_process_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_voice_voice_process_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-voice-memo":{"post":{"summary":"Process Voice Memo","operationId":"process_voice_memo_process_voice_memo_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_voice_memo_process_voice_memo_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/import-file":{"post":{"summary":"Import File","operationId":"import_file_import_file_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_file_import_file_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-pdf":{"post":{"summary":"Process Pdf","operationId":"process_pdf_process_pdf_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_pdf_process_pdf_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-image":{"post":{"summary":"Process Image","operationId":"process_image_process_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_image_process_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-pptx":{"post":{"summary":"Process Pptx","operationId":"process_pptx_process_pptx_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_pptx_process_pptx_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-docx":{"post":{"summary":"Process Docx","operationId":"process_docx_process_docx_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_docx_process_docx_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-xlsx":{"post":{"summary":"Process Xlsx","operationId":"process_xlsx_process_xlsx_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_xlsx_process_xlsx_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-csv":{"post":{"summary":"Process Csv","operationId":"process_csv_process_csv_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_csv_process_csv_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process-text":{"post":{"summary":"Process Text File","operationId":"process_text_file_process_text_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_text_file_process_text_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete-file":{"post":{"summary":"Delete File","operationId":"delete_file_delete_file_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/update-tags":{"post":{"summary":"Update Tags","operationId":"update_tags_update_tags_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTagsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import-text":{"post":{"summary":"Import Text","operationId":"import_text_import_text_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextImportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/classify":{"post":{"summary":"Classify Intent","description":"ユーザーの入力意図 (Intent) をGeminiを使って分類します。\n例: \"チャットしたい\", \"検索したい\", \"保存したい\" などを判別し、適切な処理に振り分けるために使用します。","operationId":"classify_intent_classify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassifyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query":{"post":{"summary":"Query Knowledge","description":"ユーザーのクエリを受け取り、ベクトル検索で関連情報を検索し、Geminiを使って回答を生成します (RAG)。\n1. クエリをベクトル化\n2. Supabase Vectorで類似ベクトルを検索 (タグフィルタ適用)\n3. 必要に応じてPostgreSQLから全文を取得 (Long Context)\n4. 検索結果をコンテキストとしてGeminiに渡し、回答を生成","operationId":"query_knowledge_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AskRequest":{"properties":{"query":{"type":"string","title":"Query"},"userId":{"type":"string","title":"Userid"},"threadId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Threadid"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]}},"type":"object","required":["query","userId"],"title":"AskRequest"},"Body_import_file_import_file_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_import_file_import_file_post"},"Body_process_csv_process_csv_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_csv_process_csv_post"},"Body_process_docx_process_docx_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_docx_process_docx_post"},"Body_process_image_process_image_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_image_process_image_post"},"Body_process_pdf_process_pdf_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_pdf_process_pdf_post"},"Body_process_pptx_process_pptx_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_pptx_process_pptx_post"},"Body_process_text_file_process_text_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_text_file_process_text_post"},"Body_process_voice_memo_process_voice_memo_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"},"save":{"type":"boolean","title":"Save","default":true}},"type":"object","required":["file","metadata"],"title":"Body_process_voice_memo_process_voice_memo_post"},"Body_process_voice_voice_process_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_voice_voice_process_post"},"Body_process_xlsx_process_xlsx_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"metadata":{"type":"string","title":"Metadata"}},"type":"object","required":["file","metadata"],"title":"Body_process_xlsx_process_xlsx_post"},"ClassifyRequest":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"ClassifyRequest"},"DeleteRequest":{"properties":{"fileId":{"type":"string","title":"Fileid"},"userId":{"type":"string","title":"Userid"}},"type":"object","required":["fileId","userId"],"title":"DeleteRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QueryRequest":{"properties":{"query":{"type":"string","title":"Query"},"userId":{"type":"string","title":"Userid"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"userPlan":{"type":"string","title":"Userplan","default":"FREE"}},"type":"object","required":["query","userId"],"title":"QueryRequest"},"SaveVoiceRequest":{"properties":{"userId":{"type":"string","title":"Userid"},"transcript":{"type":"string","title":"Transcript"},"summary":{"type":"string","title":"Summary"},"title":{"type":"string","title":"Title"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]}},"type":"object","required":["userId","transcript","summary","title"],"title":"SaveVoiceRequest"},"TextImportRequest":{"properties":{"text":{"type":"string","title":"Text"},"userId":{"type":"string","title":"Userid"},"source":{"type":"string","title":"Source","default":"manual"},"dbId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dbid"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"}},"type":"object","required":["text","userId"],"title":"TextImportRequest"},"UpdateTagsRequest":{"properties":{"fileId":{"type":"string","title":"Fileid"},"userId":{"type":"string","title":"Userid"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"}},"type":"object","required":["fileId","userId","tags"],"title":"UpdateTagsRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}